backend/requirements-dev.txt
# Development dependencies for CI/CD
# Include production requirements
-r requirements.txt

# Code formatting
black==24.10.0
isort==5.13.2

# Linting and type checking
flake8==7.1.1
mypy==1.13.0

# Security checks
bandit==1.7.10
safety==3.2.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
coverage==7.6.7

# Pre-commit hooks
pre-commit==4.0.1
